Resolves #1388
