+++ requirements.txt
@@ -4,7 +4,7 @@ diffrax >= 0.6.0, <= 0.7.0
 equinox >=0.11.10, <=0.13.2
 h5py >= 3.0.0, <= 3.15.1
 interpax >= 0.3.3, <= 0.3.12
-interpax_fft <= 0.0.1
+interpax_fft >= 0.0.2, <= 0.0.2
 jax-finufft >= 1.1.0
 matplotlib >= 3.7.3, <= 3.10.7
 mpmath >= 1.0.0, <= 1.3.0
